test login and landing pages
